+++ CHANGELOG.md
@@ -2,7 +2,18 @@
 
 &nbsp;
 
-#### *V2.2.5 - 28/04/2018
+#### *V3.0.0 - 04/05/2018
+
+ - Added an add source file command to the CLI (Add support for multiple source files);
+ - Added the ability for MergerJS to find the merger-config file from within the working directory;
+ - Refactored the merger init CLI process;
+ - Refactored the merger configuration file model;
+ - Refactored the codebase and further modularized it;
+ - Added more styling.
+
+&nbsp;
+
+#### V2.2.5 - 28/04/2018
 
  - Refactored and fixed the timestamps module;
  - Added some styling to the console.